# Optimize ps_pkgs