- Replicates the functionality of `Get-InstalledModule` and `Get-InstalledScript`, which apply some filters to Get-Package
  - Has a slight speed increase with Modules, and a slight speed decrease with Scripts
    - I suspect it has something to do with the quantity of modules or scripts that are found, as for me no scripts are installed and some modules are
  - Largest speed increase comes from the reuse of `Get-Package`'s output

Speed (Will populate later):
- PowerShell 5:
  - ms → ms
    - 
- PowerShell 7:
  - ms → ms
    - 

Bottlenecked Speed:
- PowerShell 5:
  - 3962ms → 3771ms
    - 5.06% Faster
- PowerShell 7:
  - 5800ms → 5289ms
    - 9.66% Faster