@rashil2000 2023-01-06
On my system, on both PS 5.1 and 7.3.1, this returns the count of both modules and scripts as 0 😕

Also, can you please make the formatting changes in a separate PR?