@Carterpersall 2023-01-12
> On my system, on both PS 5.1 and 7.3.1, this returns the count of both modules and scripts as 0 😕

For some reason, I used `-contains` instead on `-like`, which meant that my wildcards weren't working. Should be fixed now

> Also, can you please make the formatting changes in a separate PR?

Yeah, I can. When I committed that, I couldn't figure out where the setting was that was making VSCode was automatically trim whitespace. For now, I've reverted the change.